+++ Store/API/CategoryApiService.cs
@@ -30,19 +30,64 @@ namespace Store.API
         }
         public string UpdateCategory(UpdateCategoryInput input)
         {
+
             var category= context.Category.Where(c => c.Id == input.Id).FirstOrDefault();
-            category.CategoryName = input.CategoryName;
-            category.Description = input.CategoryDescription;
-            category.CategoryParent=input.CategoryParrent;
+            //if (input.CreateDate!=null)
+            //{
+            //    category!.CreateDate = input.CreateDate;
+            //}
+            //else
+            //{
+            //    category.CreateDate = DateTime.Now;
+            //}
+
+            if (!string.IsNullOrEmpty(input.CategoryName))
+            {
+                category!.CategoryName = input.CategoryName;
+            }
+            else
+            {
+                category!.CategoryName = input.CategoryName;
+            }
+            if (!string.IsNullOrEmpty(input.CategoryDescription)){
+                category!.Description = input.CategoryDescription;
+            }
+            else
+            {
+                category!.Description = input.CategoryDescription;
+            }
+            if (input.CategoryParrent >=0)
+            {
+                category!.CategoryParent = input.CategoryParrent;
+            }
+            else
+            {
+                category!.CategoryParent = input.CategoryParrent;
+            }
             context.SaveChanges();
             return "دسته بندی بروزرسانی شد";
         }
         public string DeleteCategory(DeleteCategoryInput input)
         {
-            var category = context.Category.Where(c => c.Id == input.Id).FirstOrDefault();
-            context.Category.Remove(category);
-            context.SaveChanges();
-            return "دسته بندی حذف شد";
+            if (input.Id == 0)
+            {
+                throw new Exception("ورودی شما اشتباه است");
+            }
+            else
+            {
+                var category = context.Category.Where(c => c.Id == input.Id).FirstOrDefault();
+                if (category== null) {
+                    throw new Exception("دسته بندی با این مشخصات یافت نشد");
+                }
+                else
+                {
+                    context.Category.Remove(category);
+                    context.SaveChanges();
+                    return "دسته بندی حذف شد";
+                }
+            }
+          
+            
         }
