+++ Store/API/OrdersApiService.cs
@@ -1,5 +1,6 @@
 ﻿using Store.Entitis;
 using Store.Models;
+using System.Diagnostics.Eventing.Reader;
 
 namespace Store.API
 {
@@ -39,24 +40,68 @@ namespace Store.API
         {
 
             var Orders = context.Orders.Where(o => o.Id == input.Id).FirstOrDefault();
-            Orders.count = input.count;
-            Orders.FixedPrice = input.FixedPrice;
-            Orders.otal_price_of_discounts = input.otal_price_of_discounts;
-            Orders.Original_price_payable = input.Original_price_payable;
-            Orders.ShoppingCartId=input.ShoppingCartId;
-            Orders.Id = input.Id;
-            Orders.StateId = input.StateId;
-            Orders.ProductId = input.ProductId;
+            if (input.count > 0)
+            {
+                Orders!.count = input.count;
+            }
+            else{
+                Orders!.count = input.count;
+            }
+            if (input.FixedPrice > 0)
+            {
+                Orders!.FixedPrice = input.FixedPrice;
+            }
+            else
+            {
+                Orders!.FixedPrice = input.FixedPrice;
+            }
+            if (input.otal_price_of_discounts > 0)
+            {
+                Orders!.otal_price_of_discounts = input.otal_price_of_discounts;
+            }
+            else
+            {
+                Orders!.otal_price_of_discounts = input.otal_price_of_discounts;
+            }
+            if (input.ShoppingCartId > 0)
+            {
+                Orders!.ShoppingCartId = input.ShoppingCartId;
+            }
+            else
+            {
+                Orders!.ShoppingCartId = input.ShoppingCartId;
+            }
+           
+          
+            //Orders.Id = input.Id;
+            //Orders.StateId = input.StateId;
+            //Orders.ProductId = input.ProductId;
             context.SaveChanges();
             return "نقش کاربر بروزرسانی شد";
         }
 
         public string DeleteOrders(DeleteOrders input)
         {
-            var Orders = context.Orders.Where(o => o.Id ==input.Id).FirstOrDefault();
-            context.Orders.Remove(Orders);
-            context.SaveChanges();
-            return "نقش کاربر حذف شد";
+            if (input.Id == 0)
+            {
+                throw new Exception("ورودی شما خالی است");
+            }
+            else
+            {
+                var Orders = context.Orders.Where(o => o.Id == input.Id).FirstOrDefault();
+                if(Orders == null)
+                {
+                    throw new Exception("سفارشی با این مشخصات یافت نشد");
+                }
+                else
+                {
+                    context.Orders.Remove(Orders);
+                    context.SaveChanges();
+                    return "نقش کاربر حذف شد";
+                }
+            }
+            
+            
         }
     }
 }
